Added foundation for deletion of users and closed loopholes for user manipulation. Need to fix bug in server.js.

+++ server.js
@@ -255,7 +255,7 @@ app.patch("/userDatabase/password/:id", (req, res) => {
                     .catch((error) => {
                         res.status(400).send(); // 400 for bad request.
                     });
-              }
+            }
         })
         .catch((error) => {
             res.status(400).send();
@@ -265,22 +265,44 @@ app.patch("/userDatabase/password/:id", (req, res) => {
 // A DELETE route to delete a user by their id.
 app.delete("/userDatabase/:id", (req, res) => {
     const id = req.params.id;
+    const username = req.body.username;
+    const password = req.body.password;
 
     if (!ObjectID.isValid(id)) {
         res.status(404).send();
         return;
     }
 
-    User.findByIdAndRemove(id)
-        .then((user) => {
-            if (!user) {
+    Announcement.findByUserId(id)
+        .then((announcement) => {
+            if (announcement) {
                 res.status(404).send();
-            } else {
-                res.send(user);
+                return;
             }
         })
         .catch((error) => {
-            res.status(500).send(); // Server error, could not delete.
+            res.status(400).send();
+        });
+
+
+    User.findByUsernamePassword(username, password)
+        .then((user) => {
+            if (user) {
+                User.findByIdAndRemove(id)
+                    .then((user) => {
+                        if (!user) {
+                            res.status(404).send();
+                        } else{
+                            res.send(user);
+                        }
+                    })
+                    .catch((error) => {
+                        res.status(500).send();
+                    });
+            }
+        })
+        .catch((error) => {
+            res.status(400).send();
         });
 });
 /** End of user resource routes **/
